backend/requirements.txt
fastapi
uvicorn
anthropic
openai
playwright
langgraph
langchain-core
langchain-anthropic
psycopg2-binary
python-dotenv
pydantic
beautifulsoup4
requests
pytesseract
pillow
aiohttp
asyncpg
sse-starlette
sqlalchemy
alembic
twocaptcha
